SNDG-34 View that show in real time reproductions can have one advert
I created a new component and import the forms module into the system.
Also I modified the file app.module.html to show the view implemented.
The duration and ratio of the advertisement are static, this will be
modify in the future. The advertiser only input the max price that he
want to pay.